+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/IntermediateTextInputUIView.kt
@@ -671,35 +671,3 @@ private fun NSWritingDirection.directionToStr() =
         UITextLayoutDirectionDown -> "Down"
         else -> "unknown direction"
     }
-
-internal fun toSkikoKeyboardEvent(
-    event: UIPress,
-    kind: SkikoKeyboardEventKind
-): SkikoKeyboardEvent {
-    val timestamp = (event.timestamp * 1_000).toLong()
-    return SkikoKeyboardEvent(
-        SkikoKey.valueOf(event.key!!.keyCode),
-        toSkikoModifiers(event),
-        kind,
-        timestamp,
-        event
-    )
-}
-
-private fun toSkikoModifiers(event: UIPress): SkikoInputModifiers {
-    var result = 0
-    val modifiers = event.key!!.modifierFlags
-    if (modifiers and UIKeyModifierAlternate != 0L) {
-        result = result.or(SkikoInputModifiers.ALT.value)
-    }
-    if (modifiers and UIKeyModifierShift != 0L) {
-        result = result.or(SkikoInputModifiers.SHIFT.value)
-    }
-    if (modifiers and UIKeyModifierControl != 0L) {
-        result = result.or(SkikoInputModifiers.CONTROL.value)
-    }
-    if (modifiers and UIKeyModifierCommand != 0L) {
-        result = result.or(SkikoInputModifiers.META.value)
-    }
-    return SkikoInputModifiers(result)
-}
